i made the selection grid resizable!
i made the selection grid resizable regardless of the number of squares and regardless of the screen

+++ MagnetAi/src/main/java/com/example/magnetai/Component.java
@@ -4,7 +4,7 @@ package com.example.magnetai;
 import javafx.scene.shape.Shape;
 
 public interface Component{
-	int square_size = Simulation.SQUARE_SIZE;
+	double square_size = Simulation.SQUARE_SIZE;
 	
 	int getIndex();
